set up test server
