+++ server/tests/mocks/MockAuth.ts
@@ -1,0 +1,11 @@
+import { TokenPayload } from 'google-auth-library';
+
+export const validTokenPayload: TokenPayload = {
+  email: 'dti1@cornell.edu',
+  iss: '',
+  sub: '',
+  iat: 1,
+  aud: '',
+  exp: 0,
+  hd: 'cornell.edu',
+};
