+++ server/tests/mocks/MockClasses.ts
@@ -1,0 +1,76 @@
+import { Class } from 'common';
+
+export const testClasses: Class[] = [
+  {
+    _id: 'oH37S3mJ4eAsktypy',
+    classSub: 'CS',
+    classNum: '2110',
+    classTitle: 'Object-Oriented Programming and Data Structures',
+    classPrereq: [],
+    classFull: 'CS 2110 Object-Oriented Programming and Data Structures',
+    classSems: [
+      'FA14',
+      'SP15',
+      'SU15',
+      'FA15',
+      'SP16',
+      'SU16',
+      'FA16',
+      'SP17',
+      'SU17',
+      'FA17',
+      'SP18',
+      'FA18',
+      'SU18',
+      'SP19',
+      'FA19',
+      'SU19',
+    ],
+    crossList: ['q75SxmqkTFEfaJwZ3'],
+    classProfessors: [
+      'David Gries',
+      'Douglas James',
+      'Siddhartha Chaudhuri',
+      'Graeme Bailey',
+      'John Foster',
+      'Ross Tate',
+      'Michael George',
+      'Eleanor Birrell',
+      'Adrian Sampson',
+      'Natacha Crooks',
+      'Anne Bracy',
+      'Michael Clarkson',
+    ],
+    classDifficulty: 2.9,
+    classRating: undefined,
+    classWorkload: 3,
+  },
+  {
+    _id: 'newCourse1',
+    classSub: 'MORK',
+    classNum: '1110',
+    classTitle: 'Introduction to Testing',
+    classFull: 'MORK 1110: Introduction to Testing',
+    classSems: ['FA19'],
+    classProfessors: ['Gazghul Thraka'],
+    classRating: 1,
+    classWorkload: 2,
+    classDifficulty: 3,
+    classPrereq: [],
+    crossList: [],
+  },
+  {
+    _id: 'newCourse2',
+    classSub: 'MORK',
+    classNum: '2110',
+    classTitle: 'Intermediate Testing',
+    classFull: 'MORK 2110: Intermediate Testing',
+    classSems: ['SP20'],
+    classPrereq: ['newCourse1'], // the class above
+    classProfessors: ['Gazghul Thraka'],
+    classRating: 3,
+    classWorkload: 4,
+    classDifficulty: 5,
+    crossList: [],
+  },
+];
