+++ server/tests/mocks/MockProfessors.ts
@@ -1,0 +1,16 @@
+import { Professor } from 'common';
+
+export const testProfessors: Professor[] = [
+  {
+    _id: 'prof_1',
+    fullName: 'Gazghul Thraka',
+    courses: ['newCourse1', 'newCourse2'],
+    major: 'MORK',
+  },
+  {
+    _id: 'prof_2',
+    fullName: 'Jean-Luc Picard',
+    courses: [],
+    major: 'FEDN',
+  },
+];
