+++ server/tests/mocks/MockReviews.ts
@@ -1,0 +1,25 @@
+import { Review } from 'common';
+
+export const testReviews: Review[] = [
+  {
+    _id: '4Y8k7DnX3PLNdwRPr',
+    text: 'review text for cs 2110',
+    user: 'User1234',
+    difficulty: 1,
+    class: 'oH37S3mJ4eAsktypy',
+    visible: 1,
+    reported: 0,
+    likes: 2,
+    likedBy: ['user1234', 'user0'],
+  },
+  {
+    _id: '4Y8k7DnX3PLNdwRPq',
+    text: 'review text for cs 2110 number 2',
+    user: 'User1234',
+    difficulty: 1,
+    class: 'oH37S3mJ4eAsktypy',
+    visible: 1,
+    reported: 0,
+    likedBy: [],
+  },
+];
