+++ server/tests/mocks/MockStudents.ts
@@ -1,0 +1,15 @@
+import { Student } from 'common';
+
+export const testStudents: Student[] = [
+  {
+    _id: 'Irrelevant',
+    firstName: 'John',
+    lastName: 'Smith',
+    netId: 'js0',
+    affiliation: '',
+    token: '',
+    privilege: 'regular',
+    reviews: [],
+    likedReviews: [],
+  },
+];
