+++ server/tests/mocks/MockSubjects.ts
@@ -1,0 +1,19 @@
+import { Subject } from 'common';
+
+export const testSubjects: Subject[] = [
+  {
+    _id: 'newSubject1',
+    subShort: 'MORK',
+    subFull: 'Study of Angry Fungi',
+  },
+  {
+    _id: 'angry subject',
+    subShort: 'MAD',
+    subFull: 'The Study of Anger Issues',
+  },
+  {
+    _id: 'federation subject',
+    subShort: 'FEDN',
+    subFull: 'The Study of Where No Man has Gone Before!',
+  },
+];
